% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api}
\alias{api}
\title{Create littlestar api}
\usage{
api(path_apps, path_data, validate = NULL, log_level = "info")
}
\arguments{
\item{path_apps}{Path to the app directory root}

\item{path_data}{Path to the location we will store our internal
data (outside of the app tree)}

\item{validate}{Logical, indicating if validation should be done
on responses.  This should be \code{FALSE} in production
environments.  See \link[porcelain:porcelain]{porcelain::porcelain} for details}

\item{log_level}{Logging level to use. Sensible options are "off",
"info" and "all".}
}
\value{
A \link[porcelain:porcelain]{porcelain::porcelain} object. Notably this does \emph{not}
start the server
}
\description{
Create littlestar api
}
